+++ node/resolvers/directives/checkAdminAccess.ts
@@ -3,6 +3,8 @@ import { AuthenticationError, ForbiddenError } from '@vtex/api'
 import type { GraphQLField } from 'graphql'
 import { defaultFieldResolver } from 'graphql'
 
+import sendAuthMetric, { AuthMetric } from '../../utils/metrics/auth'
+
 export class CheckAdminAccess extends SchemaDirectiveVisitor {
   public visitFieldDefinition(field: GraphQLField<any, any>) {
     const { resolve = defaultFieldResolver } = field
@@ -17,6 +19,26 @@ export class CheckAdminAccess extends SchemaDirectiveVisitor {
         vtex: { adminUserAuthToken, logger },
         clients: { identity },
       } = context
+
+      // send metric to collect data about caller and tokens used in the request
+      const operation = field.astNode?.name?.value ?? context?.request?.url
+      const metric = new AuthMetric(
+        context?.vtex?.account,
+        {
+          caller: context?.request?.headers['x-vtex-caller'] as string,
+          userAgent: context?.request?.headers['user-agent'] as string,
+          forwardedHost: context?.request?.headers[
+            'x-forwarded-host'
+          ] as string,
+          hasAdminToken: !!adminUserAuthToken,
+          hasApiToken: !!context?.request?.headers['vtex-api-apptoken'],
+          hasStoreToken: false,
+          operation,
+        },
+        'CheckAdminAccess'
+      )
+
+      sendAuthMetric(context, logger, metric)
 
       let token =
         adminUserAuthToken ?? (context?.headers.vtexidclientautcookie as string)